Make the software keyboard renderer multithreaded
